Ultrasonic Distance Sensor: Try a different strategy with pulseIn()
@XCF-KHX How would you think about something like this?

This sets a timeout for pulseIn to 11ms (believe that's 200cm?), so we should never take more time than that out of the loop(). If we miss five pulses in a row, we assume the distance is out-of-range. Otherwise, we believe in our most recent reading.

This could be made into a function that returns an int, with a static variable for missed pulses.